src/test/java/com/example/helloboot/HelloControllerTest.java
package com.example.helloboot;

import org.assertj.core.api.Assertions;
import org.junit.jupiter.api.Test;

import static org.assertj.core.api.Assertions.assertThat;
import static org.assertj.core.api.Assertions.assertThatThrownBy;
import static org.junit.jupiter.api.Assertions.*;

class HelloControllerTest {

    @Test
    void helloController() {
        // 테스트 스텁
        HelloController helloController = new HelloController(name -> name); // 수동 DI, 스프링 부트가 제공하는 Bean 이 아닌 우리가 테스트용으로 만든 Bean 으로 작업
        String ret = helloController.hello("Test");

        assertThat(ret).isEqualTo("Test");
    }

    @Test
    void failsHelloController() {
        HelloController helloController = new HelloController(name->name);

        assertThatThrownBy(() -> {
            helloController.hello(null);
        }).isInstanceOf(IllegalArgumentException.class);

        assertThatThrownBy(() -> {
            helloController.hello("");
        }).isInstanceOf(IllegalArgumentException.class);


    }
}